+++ app/jobs/locking_reminder_job.rb
@@ -8,9 +8,7 @@ class LockingReminderJob < ApplicationJob
   #
   # @param [Draw] the draw to send reminders in
   def perform(draw:)
-    students = draw.students.joins(:draw_membership)
-                   .where(draw_memberships:
-                          { intent: %w(on_campus undeclared) })
+    students = draw.students_with_intent(intents: %w(on_campus undeclared))
     admins = College.current.users.admin
     (students + admins).each { |s| send_email(user: s, draw: draw) }
   end